add Readme description

+++ app/server/routers/users_router.py
@@ -1,14 +1,15 @@
 from typing import List
 
-from fastapi import APIRouter, Depends, Body
+from fastapi import APIRouter, Depends, Body, HTTPException
 from fastapi.security import OAuth2PasswordBearer, HTTPAuthorizationCredentials, HTTPBearer
 
 from sqlalchemy.ext.asyncio import AsyncSession
 from starlette import status
 
-from app.server.crud import get_all_users, delete_some_user, update_user_information, create_new_user, check_role_user
+from app.server.crud import get_all_users, delete_some_user, update_user_information, create_new_user
 from app.server.db_helper import db_helper
 from app.server.schemas import UserUpdate, CreateUser
+from app.server.utils import get_current_user
 
 users_router = APIRouter(prefix="/users", tags=['USERS'])
 
@@ -19,19 +20,24 @@ users_router = APIRouter(prefix="/users", tags=['USERS'])
                   response_description="List all users",
                   response_model=List,
                   response_model_by_alias=False)
-async def list_users(session: AsyncSession = Depends(db_helper.scoped_session_dependency)):
+async def list_users(session: AsyncSession = Depends(db_helper.scoped_session_dependency),
+                     current_user: str = Depends(get_current_user)):
     users = await get_all_users(session)
     print(users)
     return users
 
 
 @users_router.delete("/{user_id}", description="Endpoint which removes user")
-async def delete_user(user_id: int, session: AsyncSession = Depends(db_helper.scoped_session_dependency)):
+async def delete_user(user_id: int, session: AsyncSession = Depends(db_helper.scoped_session_dependency),
+                      current_user = Depends(get_current_user)):
+    if current_user.role.name == 'user':
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN,
+                            detail='Access is denied')
     delete_result = await delete_some_user(user_id, session)
     return delete_result
 
 
-@users_router.put("/{user_name}",
+@users_router.put("/{user_id}",
                   description='Endpoint which updates user data',
                   response_description="Update a user",
                   response_model=UserUpdate,
@@ -39,19 +45,26 @@ async def delete_user(user_id: int, session: AsyncSession = Depends(db_helper.sc
                   )
 async def update_user(user_id: int,
                       user: UserUpdate = Depends(),
-                      session: AsyncSession = Depends(db_helper.scoped_session_dependency)):
-    result_update = await update_user_information(user_id, user, session)
+                      session: AsyncSession = Depends(db_helper.scoped_session_dependency),
+                      current_user = Depends(get_current_user)):
+    result_update = await update_user_information(user_id, user, session,current_user)
     return result_update
 
 
 @users_router.post("/",
                    description='Endpoint that adds a new user',
                    response_description="New user",
-                   # response_model=CreateUser,
+                   response_model=UserUpdate,
                    status_code=status.HTTP_201_CREATED,
                    response_model_by_alias=False, )
 async def create_user(user: CreateUser = Body(),
-                      session: AsyncSession = Depends(db_helper.scoped_session_dependency)):
+                      session: AsyncSession = Depends(db_helper.scoped_session_dependency),
+                      current_user = Depends(get_current_user)):
+
+
+    if current_user.role.name == 'user':
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN,
+                            detail='Access is denied')
     new_user = await create_new_user(user, session)
     return new_user
 
@@ -59,12 +72,3 @@ async def create_user(user: CreateUser = Body(),
 oauth2_scheme = OAuth2PasswordBearer(tokenUrl="token")
 
 
-async def get_current_user(credentials: HTTPAuthorizationCredentials = Depends(http_bearer),
-                           session: AsyncSession = Depends(db_helper.scoped_session_dependency)):
-    check_result = await check_role_user(credentials.credentials, session)
-    print(check_result)
-
-
-@users_router.get("/protected-endpoint")
-async def protected_endpoint(current_user: str = Depends(get_current_user)):
-    return {"message": f"Доступ разрешен для пользователя {current_user}"}
